requirements.txt
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
google-cloud==0.27.0
gunicorn==19.7.1
h5py==2.7.0
Keras==2.0.7
oauth2client==4.1.2
numpy==1.13.3
Pillow==4.3.0
six==1.11.0
scikit-learn==0.19.0
tensorflow==1.3.0
Werkzeug==0.12.2
